[GTESTREWORK] Either use system gtest or pkg gtest-1.6.0
The travis-build was failing due to use of a privately
packaged gtest-package.

This change tries to use the system-installed gtest.
If not found it will try to find a gtest-package.

Also the checks in configure.ac for external programs gsl,castxml
now fail the configure-step if not on the PATH.