[CLEAN_TESTS] Patch to clean tests and add comments
The generated files, now have a small disclaimer at the
top of the file, informing about the fact that the files
have been generated. This is just a reminder so programmers
do not modify the files without being informed that this
might be dangerous.

It is assumed that git or gerrit is used by most developers
so these type of errors should be noticed or easily fixed.

In the generated test cases unneeded repetition of the gtest
include-files were removed.

Also the sha1sum of the file is now calculated and inserted
into the generated code. This way a Makefile could check
with a simple grep-sed-statement if the code is up-to-date.